[DCFS-72] change Authenticate return value to interface{}
The goal is to make the Authenticate method return a connection client
that serves a purpose of authenticating the file exchange processes with
different disks. Given the fact that the connection clients may be of
different types, the Authenticate methods should return an interface{}
type (for now). This should be changed later

Signed-off-by: Brotholomew <bartoszek.blach@gmail.com>