[DCFS-57] refactor the disk module
To avoid import cycles the disk module has been refactored.
Additionaly:
1. old unnecessary comments are pruned;
2. the 'connect' methods were deemed not useful and thus deleted.

The main goal of this refactor is that:
1. in the inheritance tree, the parental modules 'do not see'
   the inheriting modules, e.g.: nothing from the 'models' module will
   be able to distinguish between e.g. FTPDisk or OneDriveDisk;
2. the inheriting modules are allowed to 'look up' and import the
   parental modules. E.g. the GDriveDisk will import the Disk and Volume
   interface from 'models' to correctly implement all interface methods;
3. to satisfy #1 a TypeTable is defined in the parental modules,
   which is later filled in by init() methods of the appropriate
   inheriting modules, E.g. the DiskTypeTable.

The aforementioned rules could be described as this: 'the modules can't
look down in the hierarchy, however they can look sideways and up'.
Strict following of this rule will minimize the chances of an import
cycle (due to inheritance) to 0.

Signed-off-by: Brotholomew <bartoszek.blach@gmail.com>